+++ octave/generateKpiReport.m
@@ -14,15 +14,16 @@ function generateKpiReport(hFile, dataPath, reportPath, id, kpiTitle, mode, vert
     yMin = stats(2);
     yMax = stats(3);
 
+    fprintf(hFile, "\n");
     fprintf(hFile, "## %s\n", kpiTitle);
     fprintf(hFile, "\n");
     fprintf(hFile, "| KPI    |             |\n");
     fprintf(hFile, "|--------|------------:|\n");
     fprintf(hFile, "| Steps  | %11d |\n", n);
-    fprintf(hFile, "| Mean   | %11.3g |\n", stats(4));
-    fprintf(hFile, "| Sigma  | %11.3g |\n", stats(5));
-    fprintf(hFile, "| Min    | %11.3g |\n", yMin);
-    fprintf(hFile, "| Max    | %11.3g |\n", yMax);
+    fprintf(hFile, "| Mean   | %11.3G |\n", stats(4));
+    fprintf(hFile, "| Sigma  | %11.3G |\n", stats(5));
+    fprintf(hFile, "| Min    | %11.3G |\n", yMin);
+    fprintf(hFile, "| Max    | %11.3G |\n", yMax);
 
     expTrend = 0;
     if exponential != 0 && linear(3) > exponential(3)
@@ -32,12 +33,12 @@ function generateKpiReport(hFile, dataPath, reportPath, id, kpiTitle, mode, vert
 
     if (expTrend)
       fprintf(hFile, "| Trend  | Exponential |\n");
-      fprintf(hFile, "| From   | %11.3g |\n", exp(exponential(2)));
-      fprintf(hFile, "| To     | %11.3g |\n", exp(exponential(1) * n + exponential(2)));
+      fprintf(hFile, "| From   | %11.3G |\n", exp(exponential(2)));
+      fprintf(hFile, "| To     | %11.3G |\n", exp(exponential(1) * n + exponential(2)));
     else
       fprintf(hFile, "| Trend  |      Linear |\n");
-      fprintf(hFile, "| From   | %11.3g |\n", linear(2));
-      fprintf(hFile, "| To     | %11.3g |\n", linear(1) * n + linear(2));
+      fprintf(hFile, "| From   | %11.3G |\n", linear(2));
+      fprintf(hFile, "| To     | %11.3G |\n", linear(1) * n + linear(2));
     endif
 
     # Prints histogram
@@ -46,7 +47,7 @@ function generateKpiReport(hFile, dataPath, reportPath, id, kpiTitle, mode, vert
     fprintf(hFile, "|   Value    |  Count  |         %% |\n");
     fprintf(hFile, "|-----------:|--------:|----------:|\n");
     for i = 1 : size(histogram, 2);
-      fprintf(hFile, "| %10.3g | %7d | %9.3d |\n", histogram(2, i), histogram(1, i), histogram(1, i) / n);
+      fprintf(hFile, "| %10.3G | %7d | %9.3G |\n", histogram(2, i), histogram(1, i), histogram(1, i) * 100 / n);
     endfor
 
     fprintf(hFile, "\n");